+++ packages/api/src/actions/top.ts
@@ -1,6 +1,4 @@
-import createDelegatorDAG from 'src/graph-delegation/createDelegatorDAG'
-import createDelegateDAG from 'src/graph-delegation/createDelegateDAG'
-
+import createDelegateDAG from 'src/delegate-dag'
 import { DelegationAction } from 'src/types'
 
 export default function top(
@@ -15,13 +13,12 @@ export default function top(
     offset: number
   }
 ) {
-  const outDAG = createDelegatorDAG(actions, when)
-  const inDAG = createDelegateDAG(outDAG)
+  const delegateDAG = createDelegateDAG(actions, when)
 
-  return Object.keys(inDAG)
+  return Object.keys(delegateDAG)
     .map((address) => ({
       address,
-      delegatorCount: Object.keys(inDAG[address]).length,
+      delegatorCount: Object.keys(delegateDAG[address]).length,
     }))
     .sort((a, b) => (a.delegatorCount > b.delegatorCount ? -1 : 1))
     .slice(offset, offset + limit)
